+++ .travis/check-generate.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -ex
 
-go generate gonum.org/v1/netlib/blas
+go generate gonum.org/v1/netlib/blas/netlib
 go generate gonum.org/v1/netlib/lapack/lapacke
 if [ -n "$(git diff)" ]; then
 	exit 1
