+++ blas/netlib/dgemvbench_test.go
@@ -1,4 +1,4 @@
-package blas
+package netlib
 
 import (
 	"testing"
