+++ lapack/netlib/lapack.go
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package lapack provides an interface to bindings for a C LAPACK library.
-package lapack
+// Package netlib provides an interface to bindings for a C LAPACK library.
+package netlib
 
 import (
 	"math"
